+++ WebApi/App_Start/ODataConfig.cs
@@ -1,20 +1,34 @@
 ﻿namespace forCrowd.WealthEconomy.WebApi
 {
-    using System.Web.Http;
     using Facade;
     using forCrowd.WealthEconomy.WebApi.RoutingConventions;
+    using System.Web.Http;
     using System.Web.Http.OData;
-    using System.Web.Http.OData.Extensions;
-    using System.Web.Http.OData.Routing.Conventions;
-    using System.Web.Http.OData.Routing;
     using System.Web.Http.OData.Batch;
+    using System.Web.Http.OData.Extensions;
+    using System.Web.Http.OData.Query;
+    using System.Web.Http.OData.Routing;
+    using System.Web.Http.OData.Routing.Conventions;
 
     public class ODataConfig
     {
         public static void RegisterOData(HttpConfiguration config)
         {
             // Query support
-            var odataFilter = new EnableQueryAttribute() { MaxExpansionDepth = 4 };
+            var odataFilter = new EnableQueryAttribute() {
+                AllowedArithmeticOperators = AllowedArithmeticOperators.None,
+                AllowedFunctions = AllowedFunctions.None,
+                AllowedLogicalOperators = AllowedLogicalOperators.And |
+                    AllowedLogicalOperators.Equal |
+                    AllowedLogicalOperators.Or,
+                AllowedQueryOptions = AllowedQueryOptions.Expand |
+                    AllowedQueryOptions.Filter |
+                    AllowedQueryOptions.Skip |
+                    AllowedQueryOptions.Top,
+                MaxExpansionDepth = 4,
+                MaxNodeCount = 20,
+                PageSize = 100                
+            };
             config.AddODataQueryFilter(odataFilter);
 
             // Add the CompositeKeyRoutingConvention