+++ WebApi/Filters/ValidateModelAttribute.cs
@@ -1,0 +1,32 @@
+﻿namespace forCrowd.WealthEconomy.WebApi.Filters
+{
+    using System.Net;
+    using System.Net.Http;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using System.Web.Http.Controllers;
+    using System.Web.Http.Filters;
+
+    public class ValidateModelAttribute : ActionFilterAttribute
+    {
+        public override void OnActionExecuting(HttpActionContext actionContext)
+        {
+            OnActionExecutingInternal(actionContext);
+            base.OnActionExecuting(actionContext);
+        }
+
+        public override Task OnActionExecutingAsync(HttpActionContext actionContext, CancellationToken cancellationToken)
+        {
+            OnActionExecutingInternal(actionContext);
+            return base.OnActionExecutingAsync(actionContext, cancellationToken);
+        }
+
+        void OnActionExecutingInternal(HttpActionContext actionContext)
+        {
+            if (!actionContext.ModelState.IsValid)
+            {
+                actionContext.Response = actionContext.Request.CreateErrorResponse(HttpStatusCode.BadRequest, actionContext.ModelState);
+            }
+        }
+    }
+}