+++ ngClient/_system/js/app/controllers/account/AddPasswordController.js
@@ -3,25 +3,41 @@
 
     var controllerId = 'AddPasswordController';
     angular.module('main')
-        .controller(controllerId, ['dataContext', '$location', 'logger', AddPasswordController]);
+        .controller(controllerId, ['dataContext', 'logger', '$location', AddPasswordController]);
 
-    function AddPasswordController(dataContext, $location, logger) {
+    function AddPasswordController(dataContext, logger, $location) {
         logger = logger.forSource(controllerId);
 
         var vm = this;
         vm.addPassword = addPassword;
+        vm.bindingModel = {
+            Password: '',
+            ConfirmPassword: ''
+        };
         vm.cancel = cancel;
+        vm.isSaving = false;
+        vm.isSaveDisabled = isSaveDisabled;
 
         function addPassword() {
-            dataContext.addPassword(vm)
+
+            vm.isSaving = true;
+
+            dataContext.addPassword(vm.bindingModel)
                 .success(function () {
                     logger.logSuccess('Your password has been set!', null, true);
-                    $location.url('/' + vm.userName);
+                    $location.url('/_system/account');
+                })
+                .finally(function () {
+                    vm.isSaving = false;
                 });
         }
 
         function cancel() {
             $location.url('/_system/account');
+        }
+
+        function isSaveDisabled() {
+            return vm.isSaving;
         }
     }
 })();